Performance improvements. More FPS

+++ plugins/set_strips.py
@@ -10,7 +10,7 @@ class strip_animation():
 
     def emit_row(self):
         try:
-            row_arr=full((self.datastore.LED_COUNT,4),[self.datastore.strip_vals[0], self.datastore.strip_vals[1], self.datastore.strip_vals[2], self.datastore.strip_vals[3]])
+            row_arr=full((self.datastore.LED_COUNT,4),[self.datastore.strip_vals[0], self.datastore.strip_vals[1], self.datastore.strip_vals[2], self.datastore.strip_vals[3]], dtype=int)
             return row_arr
         except Exception as err:
             print(err)
